Clamp bilinear samples to cube map face
Sample locations can overflow to neighboring cube map faces, which will lead to slight artifacts if not addressed.

The best solution would be to mutate the coordinates to sample the correct neighboring faces in case of overflow, but clamping will at least mitigate the problem and is much simpler to implement.

This change is sponsored by [Higharc](https://www.higharc.com/)